% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\title{Deprecated functions in pcaExplorer}
\arguments{
\item{...}{Ignored arguments.}
}
\value{
All functions throw a warning, with a deprecation message pointing
towards its descendent (if available).
}
\description{
Functions that are on their way to the function afterlife.
Their successors are also listed.
}
\details{
The successors of these functions are likely coming after the rework that
led to the creation of the \code{mosdef} package. See more into its
documentation for more details.
}
\section{Transitioning to the mosdef framework}{

\itemize{
\item \code{\link[=topGOtable]{topGOtable()}} is now being replaced by the more flexible
\code{\link[mosdef:run_topGO]{mosdef::run_topGO()}} function
}
}

\examples{
# try(topGOtable())

}
\author{
Federico Marini
}
