% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeds.R
\name{FEDEmakeExampleDESeqDataSet}
\alias{FEDEmakeExampleDESeqDataSet}
\title{Title}
\usage{
FEDEmakeExampleDESeqDataSet(n = 1000, m = 12, betaSD = 1,
  interceptMean = 4, interceptSD = 2, betaSD_tissue = 3,
  dispMeanRel = function(x) 4/x + 0.1, sizeFactors = rep(1, m))
}
\arguments{
\item{sizeFactors}{}
}
\description{
Title
}

