% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesbiplot.R
\name{genepca}
\alias{genepca}
\title{Title}
\usage{
genepca(x, ntop, choices = c(1, 2), arrowColors = "steelblue",
  groupNames = "group", biplot = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

