% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesbiplot.R
\name{ggbiplotFede}
\alias{ggbiplotFede}
\title{Title}
\usage{
ggbiplotFede(pcobj, choices = NULL, scale = 1, pc.biplot = TRUE,
  obs.scale = 1 - scale, var.scale = scale, groups = NULL,
  ellipse = FALSE, ellipse.prob = 0.68, labels = NULL, labels.size = 3,
  alpha = 1, var.axes = TRUE, circle = FALSE, circle.prob = 0.69,
  varname.size = 4, varname.adjust = 1.5, varname.abbrev = FALSE,
  arrowColors = NULL, groupNames = "group", returnData = FALSE,
  coordEqual = FALSE, scaleArrow = 1, useRownamesAsLabels = TRUE,
  point_size = 2, annotation = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

