% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_corr.R
\name{pair_corr}
\alias{pair_corr}
\title{Pairwise correlation plot of counts}
\usage{
pair_corr(df, method = "pearson")
}
\arguments{
\item{df}{A data frame, containing the (raw/normalized/transformed) counts}

\item{method}{Character string, one of \code{pearson} (default), \code{kendall}, or
\code{spearman} as in \code{cor}}
}
\description{
Pairwise correlation plot of counts
}

