% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca2go.R
\name{quickpca2go}
\alias{quickpca2go}
\title{Title}
\usage{
quickpca2go(se, pca_ngenes = 10000, annotation = NULL,
  inputType = "ensGene", organism = "Mm", loadings_ngenes = 500,
  background_genes = NULL, scale = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}

