% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_loadings.R
\name{hi_loadings}
\alias{hi_loadings}
\title{Title}
\usage{
hi_loadings(pcaobj, whichpc = 1, topN = 10, exprTable = NULL,
  annotation = NULL, title = "Top/bottom loadings - ")
}
\arguments{
\item{title}{}
}
\description{
Title
}

