% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeds.R
\name{makeExampleDESeqDataSet_multifac}
\alias{makeExampleDESeqDataSet_multifac}
\title{Title}
\usage{
makeExampleDESeqDataSet_multifac(n = 1000, m = 12, betaSD_condition = 1,
  betaSD_tissue = 3, interceptMean = 4, interceptSD = 2,
  dispMeanRel = function(x) 4/x + 0.1, sizeFactors = rep(1, m))
}
\arguments{
\item{n}{number of rows (genes)}

\item{m}{number of columns (samples)}

\item{betaSD_condition}{the standard deviation for condition betas, i.e. beta ~ N(0,betaSD)}

\item{betaSD_tissue}{the standard deviation for tissue betas, i.e. beta ~ N(0,betaSD)}

\item{interceptMean}{the mean of the intercept betas (log2 scale)}

\item{interceptSD}{the standard deviation of the intercept betas (log2 scale)}

\item{dispMeanRel}{a function specifying the relationship of the dispersions on
\code{2^trueIntercept}}

\item{sizeFactors}{multiplicative factors for each sample}
}
\value{
a \code{\link{DESeqDataSet}} with true dispersion,
intercept for two factors (condition and tissue) and beta values in the
 metadata columns.  Note that the true betas are provided on the log2 scale.
}
\description{
Title
}
\examples{


}

