% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca2go.R
\name{pca2go}
\alias{pca2go}
\title{Title}
\usage{
pca2go(se, pca_ngenes = 10000, annotation = NULL,
  inputType = "geneSymbol", organism = "Mm", ensToGeneSymbol = FALSE,
  loadings_ngenes = 500, background_genes = NULL, scale = FALSE, ...)
}
\description{
Title
}

