% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaExplorer-pkg.R
\docType{package}
\name{pcaExplorer-pkg}
\alias{pcaExplorer-pkg}
\alias{pcaExplorer-pkg-package}
\title{pcaExplorer: analyzing time-lapse microscopy imaging, from detection to tracking}
\description{
pcaExplorer provides functionality for interactive visualization of RNA-seq datasets
based on Principal Components Analysis. The methods provided allow for quick information
extraction and effective data exploration. A Shiny application encapsulates the whole analysis.
}
\details{
pcaExplorer provides functionality for interactive visualization of RNA-seq datasets
based on Principal Components Analysis. The methods provided allow for quick information
extraction and effective data exploration. A Shiny application encapsulates the whole analysis.
}
\author{
Federico Marini \email{marinif@uni-mainz.de}, 2016

Maintainer: Federico Marini \email{marinif@uni-mainz.de}
}
