% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstDraft.R, R/pcaExplorer.R
\docType{package}
\name{pcaExplorer}
\alias{pcaExplorer}
\alias{pcaExplorer-package}
\title{Explore a dataset from a PCA perspective}
\usage{
pcaExplorer(obj = NULL, obj2 = NULL, countmatrix = NULL, coldata = NULL,
  pca2go = NULL, annotation = NULL)
}
\arguments{
\item{obj}{A}

\item{annotation}{}
}
\value{

}
\description{
Launch a Shiny App for interactive exploration of a dataset from the perspective
of Principal Components Analysis

pcaExplorer provides functionality for interactive visualization of RNA-seq datasets
based on Principal Components Analysis. The methods provided allow for quick information
extraction and effective data exploration. A Shiny application encapsulates the whole analysis.
}
\details{
pcaExplorer provides functionality for interactive visualization of RNA-seq datasets
based on Principal Components Analysis. The methods provided allow for quick information
extraction and effective data exploration. A Shiny application encapsulates the whole analysis.
}
\examples{


}
\author{
Federico Marini \email{marinif@uni-mainz.de}, 2016

Maintainer: Federico Marini \email{marinif@uni-mainz.de}
}

