% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstDraft.R
\name{pcaExplorer}
\alias{pcaExplorer}
\title{Title}
\usage{
pcaExplorer(obj = NULL, obj2 = NULL, countmatrix = NULL, coldata = NULL,
  pca2go = NULL, annotation = NULL)
}
\arguments{
\item{annotation}{}
}
\description{
Title
}

