% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaExplorer-pkg.R, R/pcaExplorer.R
\docType{package}
\name{pcaExplorer}
\alias{pcaExplorer}
\alias{pcaExplorer-package}
\title{pcaExplorer: analyzing time-lapse microscopy imaging, from detection to tracking}
\usage{
pcaExplorer(dds = NULL, rlt = NULL, countmatrix = NULL,
  coldata = NULL, pca2go = NULL, annotation = NULL)
}
\arguments{
\item{dds}{A \code{\link{DESeqDataSet}} object. If not provided, then a \code{countmatrix}
and a \code{coldata} need to be provided. If none of the above is provided, it is possible
to upload the data during the execution of the Shiny App}

\item{rlt}{A \code{\link{DESeqTransform}} object. Can be computed from the \code{dds} object
if left NULL. If none is provided, then a \code{countmatrix}
and a \code{coldata} need to be provided. If none of the above is provided, it is possible
to upload the data during the execution of the Shiny App}

\item{countmatrix}{A count matrix, with genes as rows and samples as columns. If not provided, it is possible
to upload the data during the execution of the Shiny App}

\item{coldata}{A data.frame containing the info on the covariates of each sample. If not provided, it is possible
to upload the data during the execution of the Shiny App}

\item{pca2go}{An object generated by the \code{\link{pca2go}} function, which contains
the information on enriched functional categories in the genes that show the top or bottom loadings
in each principal component of interest. If not provided, it is possible
to compute live during the execution of the Shiny App}

\item{annotation}{A \code{data.frame} object, with row.names as gene identifiers (e.g. ENSEMBL ids)
and a column, \code{gene_name}, containing e.g. HGNC-based gene symbols}
}
\value{
A Shiny App is launched for interactive data exploration
}
\description{
pcaExplorer provides functionality for interactive visualization of RNA-seq datasets
based on Principal Components Analysis. The methods provided allow for quick information
extraction and effective data exploration. A Shiny application encapsulates the whole analysis.

Launch a Shiny App for interactive exploration of a dataset from the perspective
of Principal Components Analysis
}
\details{
pcaExplorer provides functionality for interactive visualization of RNA-seq datasets
based on Principal Components Analysis. The methods provided allow for quick information
extraction and effective data exploration. A Shiny application encapsulates the whole analysis.
}
\examples{

library(airway)
data(airway)
airway
dds_airway <- DESeq2::DESeqDataSetFromMatrix(assay(airway),
                                             colData = colData(airway),
                                             design=~dex+cell)
\dontrun{
rld_airway <- DESeq2::rlogTransformation(dds_airway)

pcaExplorer(dds_airway,rld_airway)

pcaExplorer(countmatrix = counts(dds_airway), coldata = colData(dds_airway))

pcaExplorer() # and then upload count matrix, covariate matrix (and eventual annotation)
}


}
\author{
Federico Marini \email{marinif@uni-mainz.de}, 2016

Maintainer: Federico Marini \email{marinif@uni-mainz.de}
}
