% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaplot.R
\name{pcaplot}
\alias{pcaplot}
\title{Title}
\usage{
pcaplot(x, intgroup = "condition", ntop = 500, returnData = FALSE,
  title = NULL, pcX = 1, pcY = 2, text_labels = TRUE, point_size = 3)
}
\arguments{
\item{point_size}{}
}
\value{
A value
}
\description{
Title
}
\examples{
# An example
}

