% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca2go.R
\name{topGOtable}
\alias{topGOtable}
\title{Title}
\usage{
topGOtable(DEgenes, BGgenes, ontology = "BP", maxP = 0.001, desc = "",
  annot = annFUN.org, mapping = "org.Mm.eg.db", geneID = "symbol",
  topTablerows = 200, fullNamesInRows = TRUE, addGeneToTerms = TRUE,
  plotGraph = FALSE, plotNodes = 10, writeOutput = FALSE,
  outputFile = "")
}
\arguments{
\item{addGeneToTerms}{}
}
\value{
A value
}
\description{
Title
}
\examples{
# An example
}

